[REF] base_external_dbsource_sqlite: make it sqlalchemy-wide instead
It turns out that there's a lot of modules that extend this one because they're based on sqlalchemy... that are not needed at all.

Once you select the connection engine to use, you have to provide a full sqlalchemy query string. So, you can actually select sqlite as the engine, but provide `mssql+pyodbc://` if you want (for example). It just works, because sqlalchemy is itself an abstraction layer on top of db connectors.

This module should be named instead `base_external_dbsource_sqlalchemy`, but now that it's got the sqlite name, I think it's just easier to rename the parts that are exposed to the users:

- Rename the connector to SQLAlchemy.
- Rename the methods accordingly (also according to refactor on `base_external_dbsource`).
- Update instructions.

This will probably make it more clear now that modules such as `base_external_dbsource_[mssql,odbc]` are really unnecessary.

I adopted this module too.

@moduon MT-2666