[IMP] base_..assertion: optionally use private_key value as is
In some installations, especially odoo.sh, it is not possible for the Odoo admin to put
files like the private key in the filesystem. In that case the private key can be pasted
into the private key field, and the value will be used and not interpreted as a filename.