[FIX] OD-1019, base_user_role: don't reset notification_type
The self-writable Notification field on users, which allows choosing
to be notified by email or via push, actually sets or unsets a group on
the user. This would conflict with role-based management of groups: when
setting push notification, the field would be set and push notification
received but the group would immediately be reset by base_user_role
(and some UI element would not be shown) and, in some case like module
activation, the notification type would be reset to mail, the field
being recomputed from the user's groups.

This change excludes the group in question form being unset or set by
roles. Exclusions can be added by overriding a method.